test: MaxResults setter

+++ services/cloudapi-v6/constants.go
@@ -274,7 +274,7 @@ const (
 	DefaultServerCores     = 2
 	DefaultVolumeSize      = 10
 	DefaultNicLanId        = 1
-	DefaultMaxResults      = 0
+	DefaultMaxResults      = int32(0)
 	DefaultServerCPUFamily = "AMD_OPTERON"
 	DefaultListDepth       = int32(1)
 	DefaultGetDepth        = int32(0)
